Extract "baseline.profm not deterministic" workaround as a separate kotlin dsl convention plugin
